Modified getAssignedShip to read viable chapter types from config
getAssignedShip from app/User.php:534 was modified to read a list of
chapter types from the config values in the database instead of a coded
list. This allows members of new types of chapters to view their chapter
roster.

CHANGELOG.md was updated to reflect this change as well as the previous
two minor changes.